二分练习

+++ 练习/第一章 递归与递推/Acwing 1209.带分数.cpp
@@ -11,7 +11,7 @@
  bool st[N],backup[N];
  bool check(int a,int c)
  {
-     int b=n*c-a*c;
+     long long  b=n*(long long)c-a*c;
      
      if(!a||!b||!c) return false;
      memcpy(backup,st,sizeof st);
